move API interaction to route handlers, complete data fetching for table

+++ src/tables.rs
@@ -8,7 +8,7 @@ pub struct Availability {
     pub priority: i32,
     pub start: i32,
     pub end: i32,
-    pub day: i32,
+    pub day: usize,
 }
 
 #[derive(Debug, PartialEq, Eq, Serialize, Clone)]
